Move converters and enums in TournamentEndPoint
Move it in different folders and namespaces in TournamentEndPoint
