+++ RiotSharp/TournamentEndpoint/TournamentCodeDetail.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Newtonsoft.Json;
+using RiotSharp.TournamentEndpoint.Enums;
 
 namespace RiotSharp.TournamentEndpoint
 {
