+++ RiotSharp/TournamentRiotApi.cs
@@ -5,6 +5,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using RiotSharp.TournamentEndpoint.Enums;
 
 namespace RiotSharp
 {
